Add files via upload
Dupes that work on any minecraft server. Tutorials are set in the youtube channel linked.
Put the designated file to your correct version, for example - if you are running 1.20.1, download the 1.20.1 version and put it in the mods and launch.
It is fairly easy to use, as shown in the youtube videos.